+++ .travis.yml
@@ -1,12 +1,1 @@
-sudo: false
-
 language: node_js
-
-node_js:
-    -node
-
-install:
-    -npm ci
-
-script:
-    -npm test run:coveralls
